table_definition/requirements.txt
snowflake-connector-python
pandas
cryptography
openpyxl